Move tests to seperate file
